Adding maze level
So far just the terrain, water, some basic lighting, and a few
placeholder boxes. The initial plan was for the puzzles to be in the
tunnels, but with so much space to work with I think we should probably
give the open area something in addition to the camera mechanic.